+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# [4.1.6]
+* Update quill_dropdown_button.dart.
+
+# [4.1.5]
+* Add Font Size dropdown to the toolbar.
+
+# [4.1.4]
+* New borderRadius for iconTheme.
+
 # [4.1.3]
 * Fix selection handles show/hide after paste, backspace, copy.